rename QAWolfPage -> Page
- move Page.ts and helpers to /page
